+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 * Site environment (live/testing) included in site meta for use in templates
 * `Repository` high level coordinating class combining and abstracting Stores classes
 * Error handler for Turnstile to handle/ignore client side errors
+* Construction partner CDE (Common Data Environment) distribution options
 
 ### Fixed